requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
python-dotenv>=0.19.0
yt-dlp
sqlalchemy
psycopg2-binary
parsel
requests
Pillow
